Create a unit test module for datasets and move tests there
